requirements.txt
beautifulsoup4==4.13.4
fastapi==0.115.12
fuzzywuzzy==0.18.0
keras==3.10.0
matplotlib==3.10.3
nltk==3.9.1
numpy==1.26.4
opencv_contrib_python==4.11.0.86
opencv-python-headless==4.11.0.86
pandas==2.3.0
Pillow==11.2.1
pydantic==2.11.5
pytesseract==0.3.13
Requests==2.32.4
Sastrawi==1.0.1
scikit_learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
tensorflow==2.19.0
tqdm==4.67.1
uvicorn==0.29.0
jinja2==3.1.4
python-multipart
python-Levenshtein
